refactor: move snapshot_memory_to_file into struct Vm
struct Vm encapsulates the memory, so having this function live there
makes sense. It also prepares us for the future where multiple
GuestMemoryMmap objects need to be managed.
